tee-supplicant: do not 'mkdir /data/tee' on startup
'mkdir' only on the first instance the REE FS is required, because:

1. On systems that use RPMB only, this dir isn't necessary so there's
no point in creating an unused dir.

2. AOSP systems that use OP-TEE for disk encryption depends on the
tee-supplicant service to be up before properly mounting /data. If the
tee-supplicant in return depends on 'mkdir /data/tee/' before it can
start, this creates a circular dependency and causes boot to hang.
Removing this tee-supplicant startup dependency on /data is part of the
solution chain to enable the use of OP-TEE for AOSP disk encryption.

3. Most systems seem to take care of the creation of /data/tee with
the proper permission via some sort of init script anyway before
tee-supplicant is even started.

NOTE: /data is the default value assigned to TEE_FS_PARENT_PATH

Signed-off-by: Victor Chong <victor.chong@linaro.org>
Tested-by: Victor Chong <victor.chong@linaro.org> (QEMU v8)