tee-supplicant: add device instance as start parameter for service file
tee-supplicant requires a device name as a positional argument. This can
not be provided via $OPTARGS from the EnvironmentFile, since that
variable does not account for the service file being a template where
instance-specific services (i.e. @teepriv0, @teepriv1, ...) are derived
from.

Therefore, the device instance used for each templated systemd service
needs to be included directly in the ExecStart line.

Signed-off-by: Holger Assmann <h.assmann@pengutronix.de>
Acked-by: Jerome Forissier <jerome.forissier@linaro.org>